# docs/admin/admin_swagger.yaml
basePath: /api
definitions:
...
  admin.AuditConnLogIds:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
  admin.AuditFileLogIds:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  admin.DeviceGroupForm:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
...
  model.AuditConn:
    properties:
      action:
        type: string
      close_time:
        type: integer
      conn_id:
        type: integer
      created_at:
        type: string
      from_name:
        type: string
      from_peer:
        type: string
      id:
        type: integer
      ip:
        type: string
      peer_id:
        type: string
      session_id:
        type: string
      type:
        type: integer
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  model.AuditConnList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AuditConn'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  model.AuditFile:
    properties:
      created_at:
        type: string
      from_name:
        type: string
      from_peer:
        type: string
      id:
        type: integer
      info:
        type: string
      ip:
        type: string
      is_file:
        type: boolean
      num:
        type: integer
      path:
        type: string
      peer_id:
        type: string
      type:
        type: integer
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  model.AuditFileList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AuditFile'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  model.DeviceGroup:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.Group:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        type: integer
      updated_at:
        type: string
    type: object
  model.GroupList:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Group'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: RustDesk API Server
  title: RustDesk API
  version: "1.0"
paths:
...
  /admin/audit_conn/batchDelete:
    post:
      consumes:
      - application/json
      description: Suppression par lot des journaux de connexion
      parameters:
      - description: Journaux de connexion
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.AuditConnLogIds'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Suppression par lot des journaux de connexion
      tags:
      - Journaux de connexion
  /admin/audit_conn/delete:
    post:
      consumes:
      - application/json
      description: Supprimer un journal de connexion
      parameters:
      - description: Informations sur le journal de connexion
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AuditConn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Supprimer un journal de connexion
      tags:
      - Journaux de connexion
  /admin/audit_conn/list:
    get:
      consumes:
      - application/json
      description: Liste des journaux de connexion
      parameters:
      - description: Numéro de page
        in: query
        name: page
        type: integer
      - description: Taille de la page
        in: query
        name: page_size
        type: integer
      - description: Appareil cible
        in: query
        name: peer_id
        type: integer
      - description: Appareil source
        in: query
        name: from_peer
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AuditConnList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Liste des journaux de connexion
      tags:
      - Journaux de connexion
  /admin/audit_file/batchDelete:
    post:
      consumes:
      - application/json
      description: Suppression par lot des journaux de fichiers
      parameters:
      - description: Journaux de fichiers
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.AuditFileLogIds'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Suppression par lot des journaux de fichiers
      tags:
      - Journaux de fichiers
  /admin/audit_file/delete:
    post:
      consumes:
      - application/json
      description: Supprimer un journal de fichier
      parameters:
      - description: Informations sur le journal de fichier
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AuditFile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Supprimer un journal de fichier
      tags:
      - Journaux de fichiers
  /admin/audit_file/list:
    get:
      consumes:
      - application/json
      description: Liste des journaux de fichiers
      parameters:
      - description: Numéro de page
        in: query
        name: page
        type: integer
      - description: Taille de la page
        in: query
        name: page_size
        type: integer
      - description: Appareil cible
        in: query
        name: peer_id
        type: integer
      - description: Appareil source
        in: query
        name: from_peer
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.AuditFileList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Liste des journaux de fichiers
      tags:
      - Journaux de fichiers
...
  /admin/device_group/create:
    post:
      consumes:
      - application/json
      description: Créer un groupe de périphériques
      parameters:
      - description: Informations sur le groupe de périphériques
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.DeviceGroupForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.DeviceGroup'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Créer un groupe de périphériques
      tags:
      - Groupe de périphériques
  /admin/device_group/delete:
    post:
      consumes:
      - application/json
      description: Supprimer le groupe de périphériques
      parameters:
      - description: Informations sur le groupe
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.DeviceGroupForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Supprimer le groupe de périphériques
      tags:
      - Groupe de périphériques
  /admin/device_group/detail/{id}:
    get:
      consumes:
      - application/json
      description: Détails du groupe de périphériques
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Group'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Détails du groupe de périphériques
      tags:
      - Groupe de périphériques
  /admin/device_group/list:
    get:
      consumes:
      - application/json
      description: Liste des groupes
      parameters:
      - description: Numéro de page
        in: query
        name: page
        type: integer
      - description: Taille de la page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.GroupList'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Liste des groupes
      tags:
      - Groupe
  /admin/device_group/update:
    post:
      consumes:
      - application/json
      description: Modifier le groupe de périphériques
      parameters:
      - description: Informations sur le groupe
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.DeviceGroupForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Group'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - token: []
      summary: Modifier le groupe de périphériques
      tags:
      - Groupe de périphériques
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  token:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"